+++ src/features/coal/practical-playgrounds/AddressingModePlaygroundPage.jsx
@@ -1,14 +1,14 @@
 import React from "react";
 import { Link } from "react-router-dom";
 import { ArrowLeft } from "lucide-react";
-import Navbar from "../../shared/components/navbar";
-import Footer from "../../shared/components/Footer";
-import { useTheme } from "../../shared/context/ThemeContext";
-import usePointerGlow from "../../shared/hooks/usePointerGlow";
-import { coalCourseMeta } from "../../shared/data/coalCourseOutline";
-import AddressingModePlayground from "./components/AddressingModePlayground";
-import "../home/Home.css";
-import "../../shared/styles/LearningResourcesPage.css";
+import Navbar from "../../../shared/components/navbar";
+import Footer from "../../../shared/components/Footer";
+import { useTheme } from "../../../shared/context/ThemeContext";
+import usePointerGlow from "../../../shared/hooks/usePointerGlow";
+import { coalCourseMeta } from "../../../shared/data/coalCourseOutline";
+import AddressingModePlayground from "../components/AddressingModePlayground";
+import "../../home/Home.css";
+import "../../../shared/styles/LearningResourcesPage.css";
 
 const COAL_ACCENT = coalCourseMeta.accent;
