+++ src/features/coal/styles/AluFlagsSimulator.css
