+++ src/features/coal/styles/AssemblyStackSimulator.css
