+++ src/features/coal/styles/CPUExplorer.css
