+++ src/features/coal/styles/CoalLayout.css
