+++ src/features/coal/styles/CoalPages.css
