+++ src/features/coal/styles/CoalSidebar.css
