+++ src/features/coal/styles/InstructionLaboratoryPage.css
